Обход сталкерами физ. объектов и открывание/закрывание дверей из ЗП
Восстановил консистентность

Раньше, когда маска была u64, по умолчанию она была равна -1, т.е. все
биты установлены. Потом добавили еще u64 и при создании SPHBonesData
там правильно устанавливались все биты. А вот тут вот, зачем-то
обнулялись.

Экспортировал set_ai_obstacle() и is_door_registered_for_npc()

Первое переключает возвращаемое значения метода is_ai_obstacle() у
физ. объектов. Т.е. нужно или не нужно обходить этот объект.

is_door_registered_for_npc() возвращает зарегистрирован этот объект в
менеджере дверей или нет.

В register_door_for_npc() заменил вылет на вывод в лог в случае, если
движок не может зарегистрировать эту дверь. Например на посту Долга на
Свалке, перед переходом в Бар, есть откатывающиеся ворота. Их движок
не может зарегистрировать из-за object.get_door_vectors().

В ph_door.script нужно проверять is_door_registered_for_npc() перед
вызовами методов дверей.

Поменял LL_VisibleBoneCount() без visimask.count()

Во первых, каким-то образом visimask.count() вернуло 65 видимых
костей, при том, что всего костей было 6.

Во вторых, внутри себя visimask.count() делает цикл из 128
итераций. Лучше сделать цикл по кол-ву костей, а не бит в маске. В
этом случае итераций будет меньше.

Теперь результат LL_VisibleBoneCount() всегда будет соответствовать
кол-ву костей.

Если посмотреть на оригинальную реализацию, когда маска была u64, то
там учитывались только биты относящиеся к имеющимся костям, а не все.

Автоопределение дверей, для отключения их обхода

В ЗП двери сделаны на секции

[physic_door]
is_is_obstacle = false

в которой указано отключение обхода препятствий. В ТЧ отдельная секция
для дверей не используется, а их обход нужно отключать. Автоугадав
будет отключать обход для физ. объектов, у которых есть кость с именем
"door".

+++ ogsr_engine/Layers/xrRender/DetailManager_Decompress.cpp
@@ -93,7 +93,7 @@ void		CDetailManager::cache_Decompress(Slot* S)
 #else
 	xrc.box_options		(CDB::OPT_FULL_TEST); 
 	xrc.box_query		(g_pGameLevel->ObjectSpace.GetStaticModel(),bC,bD);
-	u32	triCount		= xrc.r_count	();
+	size_t	triCount	= xrc.r_count	();
 	CDB::TRI*	tris	= g_pGameLevel->ObjectSpace.GetStaticTris();
 	Fvector*	verts	= g_pGameLevel->ObjectSpace.GetStaticVerts();
 #endif
